renamed a few files to match convention
