+++ client/src/components/Review/ReviewWidget.jsx
@@ -64,6 +64,7 @@ function ReviewWidget () {
         <Parent>
           <ReviewSummary  filterReviews={addFilters}/>
           <ReviewList reviews={reviews}/>
+          <ReviewButtons/>
         </Parent>
       </>
     )
